Added an ID config and fixed world gen and fixed using fix id´s in some places instead of variables

+++ common/bb/mods/bbbc/block/Faceblock.java
@@ -1,6 +1,7 @@
 package bb.mods.bbbc.block;
 
 import bb.mods.bbbc.lib.Block_Names;
+import bb.mods.bbbc.lib.LoadedIDs;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.material.Material;
 import net.minecraft.tileentity.TileEntity;
@@ -11,18 +12,16 @@ public class Faceblock extends BlockContainer {
 
 	public static final int shiftedIndex = 165;
 
+	public Faceblock(int id) {
 
-	public Faceblock(int id, Material material) {
-
-		super(id, material);
+		super(id, Material.wood);
 
 		setUnlocalizedName(Block_Names.FACEBLOCK);
 
 	}
-	
-	public TileEntity createTileEntity(World world, int metadata)
-	{
-	   return new TileEntityFace();
+
+	public TileEntity createTileEntity(World world, int metadata) {
+		return new TileEntityFace();
 	}
 
 	public void breakBlock(World world, int x, int y, int z, int p, int t) {
@@ -31,12 +30,17 @@ public class Faceblock extends BlockContainer {
 				y, z);
 
 		if (tileEntiy != null) {
-			world.destroyBlock(tileEntiy.primary_x, tileEntiy.primary_y,
-					tileEntiy.primary_z, false);
+			if (world.getBlockId(tileEntiy.primary_x, tileEntiy.primary_y,
+					tileEntiy.primary_z) == LoadedIDs.Block_BigBlock)
+				world.destroyBlock(tileEntiy.primary_x, tileEntiy.primary_y,
+						tileEntiy.primary_z, false);
 
-			world.removeBlockTileEntity(tileEntiy.primary_x,
-					tileEntiy.primary_y, tileEntiy.primary_z);
+			//world.removeBlockTileEntity(tileEntiy.primary_x,
+				//	tileEntiy.primary_y, tileEntiy.primary_z);
+			
 		}
+		
+		world.removeBlockTileEntity(x, y, z);
 
 	}
 
@@ -44,30 +48,28 @@ public class Faceblock extends BlockContainer {
 	public void onNeighborBlockChange(World world, int x, int y, int z, int p) {
 		TileEntityFace tileEntiy = (TileEntityFace) world.getBlockTileEntity(x,
 				y, z);
-		
-		if(tileEntiy != null){
-			if(world.getBlockId(tileEntiy.primary_x,tileEntiy.primary_y,tileEntiy.primary_z)<1){
-				
+
+		if (tileEntiy != null) {
+			if (world.getBlockId(tileEntiy.primary_x, tileEntiy.primary_y,
+					tileEntiy.primary_z) < 1||world.getBlockId(tileEntiy.primary_x, tileEntiy.primary_y,
+							tileEntiy.primary_z)!=LoadedIDs.Block_BigBlock) {
+
 				world.destroyBlock(x, y, z, false);
 				world.removeBlockTileEntity(x, y, z);
-				
+
 			}
-		}
-		else{
+		} else {
 			world.destroyBlock(x, y, z, false);
 			world.removeBlockTileEntity(x, y, z);
-			
+
 		}
-		
 
 	}
-		
+
 	@Override
-	public boolean isOpaqueCube(){
+	public boolean isOpaqueCube() {
 		return false;
 	}
-	
-	
 
 	@Override
 	public TileEntity createNewTileEntity(World world) {
